package/webkitgtk: bump to version 2.26.0
This is a new major release which brings in many improvements and new
features. For a complete list, please refer to the release notes:

  https://webkitgtk.org/2019/09/09/webkitgtk2.26.0-released.html

A small patch is added which fixes a build failure when X11 headers
are not available (for example, when building a Wayland-only system)

The new support for the WPE renderer on Wayland and the new Bubblewrap
sandbox need additional dependencies and therefore are explicitly
disabled at the moment.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 38b740ec6f44b4ba561f2d8291e58c45a10a0423)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>