docs/manual/quickstart: update output directory contents documentation
Update the documentation for the output/host/ directory to mention
that it contains the sysroot for the target toolchain, as well as the
host tools required for running buildroot.

Update the staging/ documentation to reflect that it is a link to the
target toolchain sysroot in the host/ directory.

Signed-off-by: Michael Drake <michael.drake@codethink.co.uk>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
(cherry picked from commit 728d6800bb0af0fcbd66db701412b1ad0572a5dc)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>