package/eudev: Fix monitor starting for kernels w/o CONFIG_SHMEM
When the kernel has CONFIG_SHMEM disabled, /dev is a ramfs (instead of a
tmpfs) and the name_to_handle_at system call is not supported. This
causes eudev's monitor application to exit on startup.

Upstream eudev has added this fix which is not yet part of a release.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
(cherry picked from commit 1b81eb7d04f161cdee7cd59d4443e84401ac88ef)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>